services/ingestion/doc_ingester/requirements.txt
psycopg[binary]>=3.2
pydantic>=2.8
pydantic-settings>=2.4
typer>=0.12
pypdf>=5.0
python-docx>=1.1
defusedxml>=0.7
